requirements.txt
numpy
pillow
cython
git+https://github.com/lucasb-eyer/pydensecrf.git
torch
torchvision
tqdm
opencv-python